+++ frontend/kezuler-fe/src/components/main-page/overview-modal/OverviewTextarea.tsx
@@ -1,11 +1,10 @@
-import React, { useEffect, useState } from 'react';
-import { isIOS, isMobile } from 'react-device-detect';
+import React from 'react';
 import { FieldError, FieldPath, UseFormRegisterReturn } from 'react-hook-form';
 import TextareaAutosize from '@mui/material/TextareaAutosize';
 import classNames from 'classnames';
 
+import useIOSScroll from 'src/hooks/useIOSScroll';
 import { OverviewEventForm } from 'src/types/Overview';
-import { focusDisable, focusEnable } from 'src/utils/iosScrollDisable';
 
 interface Props {
   textareaClassName: string;
@@ -24,19 +23,7 @@ function OverviewTextarea({
   defaultValue,
   allowNewLine,
 }: Props) {
-  const [focused, setFocused] = useState(false);
-  const onFocus = () => setFocused(true);
-  const onBlur = () => setFocused(false);
-
-  // useEffect(() => {
-  //   if (isMobile && isIOS) {
-  //     if (focused) {
-  //       focusDisable();
-  //     } else {
-  //       focusEnable();
-  //     }
-  //   }
-  // }, [focused]);
+  useIOSScroll();
 
   const handleKeyDown = (e: React.KeyboardEvent) => {
     if (!allowNewLine && e.code === 'Enter') {
@@ -54,8 +41,6 @@ function OverviewTextarea({
         defaultValue={defaultValue}
         {...registered}
         placeholder={placeholder}
-        onFocus={onFocus}
-        onBlur={onBlur}
       />
       {error && <div className={'overview-error-text'}>{error.message}</div>}
     </div>
